refactor : 타입 파일 선언
